+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
   * Added Digraph#ith_vertex and Digraph#ith_edge 
   * Added Digraph#vertex_count and Digraph#edge_count
+  * Added Digraph.random to generate graph for tests and benchmarks
 
 * Internals & Devel